requirements.txt
Django==1.11
djangorestframework==3.6.2
docutils==0.14
pytz==2017.3
